DJANGO - PLATAFORMA GESTION DE INVENTARIO PARA UN ALMACEN/requirements.txt
Django==4.2.8
python-barcode==0.16.1
pillow==11.1.0
xlsxwriter==3.1.9
